+++ acslib/__init__.py
@@ -3,3 +3,5 @@
 __author__ = """Jeremy Gibson"""
 __email__ = "jmgibso3@ncsu.edu"
 __version__ = "0.1.0"
+
+from acslib.ccure.crud import CcureAPI
